updated - and moved classes related to the widgets class
